EXPAND ALL THE BITS!

+++ src/map/ability.cpp
@@ -208,7 +208,7 @@ void CAbility::setRecastId(uint16 recastId)
     m_recastId = recastId;
 }
 
-void CAbility::setCE(uint16 CE)
+void CAbility::setCE(uint32 CE)
 {
     m_CE = CE;
 }
@@ -218,7 +218,7 @@ uint16 CAbility::getCE()
     return m_CE;
 }
 
-void CAbility::setVE(uint16 VE)
+void CAbility::setVE(uint32 VE)
 {
     m_VE = VE;
 }
