+++ src/map/utils/battleutils.h
@@ -170,7 +170,7 @@ namespace battleutils
     bool				isValidSelfTargetWeaponskill(int wsid);
     int16				CalculateBaseTP(int delay);
     void				GenerateCureEnmity(CCharEntity* PSource, CBattleEntity* PTarget, uint16 amount);
-    void                GenerateInRangeEnmity(CBattleEntity* PSource, int16 CE, int16 VE);
+    void                GenerateInRangeEnmity(CBattleEntity* PSource, int32 CE, int32 VE);
 
     CItemWeapon*		GetEntityWeapon(CBattleEntity* PEntity, SLOTTYPE Slot);
     CItemArmor*			GetEntityArmor(CBattleEntity* PEntity, SLOTTYPE Slot);
